// frontend/src/api-generated/models/nhaCungCapResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface NhaCungCapResponse {
  id?: number;
  maNhaCungCap?: string;
  idDonVi?: number;
  tenNhaCungCap?: string;
  maSoThue?: string;
  nguoiLienHe?: string;
  soDienThoai?: string;
  email?: string;
  diaChi?: string;
  ghiChu?: string;
  trangThai?: string;
  thoiGianTao?: string;
  thoiGianCapNhat?: string;
}