// frontend/src/api-generated/models/nhaCungCapRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface NhaCungCapRequest {
  /**
     * @minLength 0
     * @maxLength 50
     */
  maNhaCungCap?: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  tenNhaCungCap: string;
  /**
     * @minLength 0
     * @maxLength 20
     */
  maSoThue?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  nguoiLienHe?: string;
  /**
     * @minLength 0
     * @maxLength 20
     */
  soDienThoai?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  email?: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  diaChi?: string;
  ghiChu?: string;
}